refactor: modularize extractfacts, update import paths
Updated import paths in test files to reflect new module structure after extractfacts functionality was refactored into separate modules (document_reader, single_extractor, core). This change ensures tests continue to work with the new modular organization of the codebase.
