+++ litassist/commands/extractfacts/__init__.py
@@ -1,0 +1,4 @@
+"""Extract facts command package."""
+from .core import extractfacts
+
+__all__ = ["extractfacts"]
