+++ tests/unit/test_noverify_flag.py
@@ -29,11 +29,11 @@ class TestNoVerifyFlag:
         self.mock_client.validate_citations.return_value = []
         self.mock_client.model = "test-model"
 
-    @patch("litassist.commands.extractfacts.get_config")
-    @patch("litassist.commands.extractfacts.LLMClientFactory.for_command")
-    @patch("litassist.commands.extractfacts.verify_content_if_needed")
-    @patch("litassist.commands.extractfacts.save_command_output")
-    @patch("litassist.commands.extractfacts.PROMPTS")
+    @patch("litassist.commands.extractfacts.document_reader.get_config")
+    @patch("litassist.commands.extractfacts.core.LLMClientFactory.for_command")
+    @patch("litassist.commands.extractfacts.core.verify_content_if_needed")
+    @patch("litassist.commands.extractfacts.core.save_command_output")
+    @patch("litassist.commands.extractfacts.single_extractor.PROMPTS")
     def test_extractfacts_with_noverify_skips_verification(
         self, mock_prompts, mock_save, mock_verify, mock_factory, mock_config
     ):
@@ -67,11 +67,11 @@ class TestNoVerifyFlag:
         finally:
             Path(test_file).unlink()
 
-    @patch("litassist.commands.extractfacts.get_config")
-    @patch("litassist.commands.extractfacts.LLMClientFactory.for_command")
-    @patch("litassist.commands.extractfacts.verify_content_if_needed")
-    @patch("litassist.commands.extractfacts.save_command_output")
-    @patch("litassist.commands.extractfacts.PROMPTS")
+    @patch("litassist.commands.extractfacts.document_reader.get_config")
+    @patch("litassist.commands.extractfacts.core.LLMClientFactory.for_command")
+    @patch("litassist.commands.extractfacts.core.verify_content_if_needed")
+    @patch("litassist.commands.extractfacts.core.save_command_output")
+    @patch("litassist.commands.extractfacts.single_extractor.PROMPTS")
     def test_extractfacts_without_noverify_uses_verification(
         self, mock_prompts, mock_save, mock_verify, mock_factory, mock_config
     ):
@@ -236,11 +236,11 @@ class TestVerificationDefaults:
         self.mock_client.verify.return_value = ("Content", "model")
         self.mock_client.validate_citations.return_value = []
 
-    @patch("litassist.commands.extractfacts.get_config")
-    @patch("litassist.commands.extractfacts.LLMClientFactory.for_command")
-    @patch("litassist.commands.extractfacts.verify_content_if_needed")
-    @patch("litassist.commands.extractfacts.save_command_output")
-    @patch("litassist.commands.extractfacts.PROMPTS")
+    @patch("litassist.commands.extractfacts.document_reader.get_config")
+    @patch("litassist.commands.extractfacts.core.LLMClientFactory.for_command")
+    @patch("litassist.commands.extractfacts.core.verify_content_if_needed")
+    @patch("litassist.commands.extractfacts.core.save_command_output")
+    @patch("litassist.commands.extractfacts.single_extractor.PROMPTS")
     def test_extractfacts_default_enables_verification(
         self, mock_prompts, mock_save, mock_verify, mock_factory, mock_config
     ):
